requirements.txt
fastapi 
uvicorn 
python-multipart

python-jose[cryptography] 
passlib[bcrypt]
pyjwt

psycopg2-binary 
sqlalchemy 

pydantic[email]

pymupdf
docx2txt
python-docx
spacy

sentence-transformers

python-dateutil